+++ src/Transform.cpp
@@ -10,7 +10,7 @@
 Transform::Transform() {
 }
 
-Mat4f Transform::scale(const float sx, const float sy, const float sz) {
+Mat4f Transform::scale(const double sx, const double sy, const double sz) {
 	Mat4f ret;
 	ret[0].x = sx;
 	ret[1].y = sy;
@@ -18,7 +18,7 @@ Mat4f Transform::scale(const float sx, const float sy, const float sz) {
 	return ret;
 }
 
-Mat4f Transform::translate(const float tx, const float ty, const float tz) {
+Mat4f Transform::translate(const double tx, const double ty, const double tz) {
 	Mat4f ret;
 	ret[3].x = tx;
 	ret[3].y = ty;
@@ -26,8 +26,8 @@ Mat4f Transform::translate(const float tx, const float ty, const float tz) {
 	return ret;
 }
 
-Mat4f Transform::rotate(const float rx, const float ry, const float rz,
-		float degrees) {
+Mat4f Transform::rotate(const double rx, const double ry, const double rz,
+		double degrees) {
 	//std::cout << rx <<", " << ry << ", " << rz<< ", " << degrees << std::endl;
 	Vec3f axis(rx, ry, rz);
 	Mat4f ret;
@@ -40,18 +40,18 @@ Mat4f Transform::rotate(const float rx, const float ry, const float rz,
 			0,0,0,1);
 
 	Mat4f secondPartMat = Mat4f(
-			0.0f, axis[2], -axis[1],0.0f,
-			-axis[2], 0.0f, axis[0],0.0f,
-			axis[1], -axis[0], 0.0f,0.0f,
-			0.0f,0.0f,0.0f,1.0f);
+			0.0, axis[2], -axis[1],0.0,
+			-axis[2], 0.0, axis[0],0.0,
+			axis[1], -axis[0], 0.0,0.0,
+			0.0,0.0,0.0,1.0);
 
-	float radiants = degrees * M_PI / 180;
+	double radiants = degrees * M_PI / 180;
 
-	ret = ((float)cos(radiants)) * identityMat;
-	ret = ret + ((float)(1 - cos(radiants))) * firstPartMat;
-	ret = ret + ((float)sin(radiants)) * secondPartMat;
+	ret = ((double)cos(radiants)) * identityMat;
+	ret = ret + ((double)(1 - cos(radiants))) * firstPartMat;
+	ret = ret + ((double)sin(radiants)) * secondPartMat;
 	//now we should reset the last element
-	ret[3][3]=1.0f;
+	ret[3][3]=1.0;
 	return ret;
 
 }
